+++ matchMultipleTemplate/geneticAlgorithm/actualStiches/teste1.png
