+++ matchMultipleTemplate/geneticAlgorithm/actualStiches/teste1_original.png
