+++ matchMultipleTemplate/geneticAlgorithm/actualStiches/teste2.png
