+++ matchMultipleTemplate/geneticAlgorithm/actualStiches/teste3.png
